Organize configuration files and scripts
